pyproject.toml
[tool.poetry]
name = "ai-agent-template"
version = "0.1.0"
description = "REPLACE WITH PROJECT DESCRIPTION"
authors = ["REPLACE WITH YOUR NAME <REPLACE WITH YOUR EMAIL>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3"
langchain-community = "^0.3"
langchain-openai = "^0.2"
openai = "^1.0"
fastapi = "^0.111"
uvicorn = {extras = ["standard"], version = "^0.30"}
psycopg2-binary = "^2.9"
sqlalchemy = "^2.0"
supabase = "^2.0"
python-dotenv = "^1.0"
pydantic = "^2.0"
apscheduler = "^3.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
black = "^24.0"
ruff = "^0.4"
httpx = "^0.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
